+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 - PR #5843 Add `filters` parameter to Python `read_parquet` function for filtering row groups
 - PR #6116 Add `filters` parameter to Python `read_orc` function or filtering
 - PR #5974 Use libcudf instead of cupy for `arange` or column creation from a scalar.
+- PR #5874 Add `COLLECT` groupby aggregation
 - PR #6119 Add support for `dayofweek` property in `DateTimeIndex` and `DatetimeProperties`
+- PR #6125 Add support for `Series.mode` and `DataFrame.mode`
+- PR #6142 Add Python `read_orc_statistics` function for reading file- and stripe-level statistics
 
 ## Improvements